Separated ramp calculations
